Odio mi vida commit
Quiero hacer nieves perdon

+++ src/app/libros/pages/libro-pages/libro-pages.component.ts
@@ -3,58 +3,40 @@ import { Component, Input, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { LibroService } from '../../../services/libro.service';
 import { switchMap } from 'rxjs';
-// import { ActivatedRoute } from '@angular/router';
-// import { Router } from 'express';
-// import { LibroService } from '../../../services/libro.service';
 
 @Component({
   selector: 'app-libro-pages',
   templateUrl: './libro-pages.component.html',
-  styles: `
-            .img-card-libro{
-              max-height: 70vh;
-            }
-            .papi{
-
-            }
-          `
+  styles: ``
 })
 export class LibroPagesComponent implements OnInit{
   @Input()
   public libro!: Libro;
 
   constructor(
-              private activatedRoute : ActivatedRoute,
-              private router : Router,
-              private libroService: LibroService
-              ){}
+    private activatedRoute : ActivatedRoute,
+    private libroService: LibroService,
+    private router : Router,
+  ){
+    this.activatedRoute.params.pipe(switchMap( params => this.libroService.getByID(params['isbn'])))
+    .subscribe( resp =>{this.libro= resp[0];})
+  }
 
   ngOnInit(): void {
-    // Paso 1: Coger la id de la querystring
-    this.activatedRoute.params
-        .pipe(
-          // Le damos la id ya al service
-          switchMap( params => this.libroService.getByID(params['isbn']))
-        )
-        .subscribe( resp =>{
-          // console.log(resp[0].Tema);
-
-          this.libro= resp[0];
-
-          // console.log(this.libro.Tema.Tipo);
-        }
-        )
-
   }
 
   public borrarLibro(){
-    this.libroService.delete(this.libro.ISBN).subscribe(resp=>{
+    console.log(this.libro)
+    if(confirm(`Estás seguro que quieres eliminar [ ${this.libro.Titulo} ]`)){
+      this.libroService.delete(this.libro.ISBN).subscribe(resp=>{
 
-      // TODO que una ventana o mensajeme diga que el libro se borro
-      console.log(resp);
+        if(resp) alert("Libro eliminado")
 
+        console.log(resp)
+        this.router.navigate(["../"])
+      }
+      );
     }
-    );
   }
